# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Полная документация API для электронного дневника",
        "title": "API системы электронного дневника",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/disciplines": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplines"
                ],
                "summary": "Получить список дисциплин с фильтрацией",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID преподавателя",
                        "name": "teacher_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "student_group_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Ограничение",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service_internal_domain_models.Discipline"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplines"
                ],
                "summary": "Создать дисциплину",
                "parameters": [
                    {
                        "description": "Дисциплина",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.Discipline"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.Discipline"
                        }
                    }
                }
            }
        },
        "/api/v1/disciplines/public": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplines"
                ],
                "summary": "Получить список публичных дисциплин",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID преподавателя",
                        "name": "teacher_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID группы студентов",
                        "name": "student_group_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID учебного года",
                        "name": "academic_year_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Ограничение",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service_internal_domain_models.DisciplinePublic"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/disciplines/public/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplines"
                ],
                "summary": "Получить публичную дисциплину по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID дисциплины",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.DisciplinePublic"
                        }
                    }
                }
            }
        },
        "/api/v1/disciplines/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplines"
                ],
                "summary": "Получить дисциплину по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID дисциплины",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.Discipline"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplines"
                ],
                "summary": "Обновить дисциплину",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID дисциплины",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Дисциплина",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.Discipline"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.Discipline"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplines"
                ],
                "summary": "Удалить дисциплину",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID дисциплины",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service_internal_domain_models.Discipline": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "discipline_id": {
                    "type": "integer"
                },
                "discipline_name": {
                    "type": "string"
                },
                "student_group_id": {
                    "type": "integer"
                },
                "teacher_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "service_internal_domain_models.DisciplinePublic": {
            "type": "object",
            "properties": {
                "academic_year_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "curator_first_name": {
                    "type": "string"
                },
                "curator_id": {
                    "type": "integer"
                },
                "curator_last_name": {
                    "type": "string"
                },
                "curator_middle_name": {
                    "type": "string"
                },
                "discipline_id": {
                    "type": "integer"
                },
                "discipline_name": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "student_group_id": {
                    "type": "integer"
                },
                "student_group_name": {
                    "type": "string"
                },
                "teacher_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}